fix oob read in value_get, signal races, stale builds [AI]
memory safety:
- procfs: rewrite value_get bounds; old length math underflowed and
  could read past the buffer (regression tested in test-edge-cases)
- procfs/obs: size pid path buffers sizeof*3 instead of *8
- tests/test-edge-cases: fix 6 unsigned int* -> unsigned short*
  interval pointer mismatches

correctness:
- main: clear signal mask atomically (__sync_fetch_and_and); a SIGHUP
  landing between read and clear could drop pending rt signals
- main: define G_SIGNAL_MAX after HAVE_RT_SIGNALS; was always 31 while
  the rt range is 30, so .signal=31 passed b_init then aborted startup
- gpu-nvidia: fix nvml fallback build (buf.dev -> dev), print the real
  nvml error string, null b_gpu.buf after free
- obs.h: sync b_write_obs prototype with definition (9 vs 10 params)
- cpu: guard div-by-zero on same-jiffy samples and same-tick power reads
- time: check localtime return; fix Agu->Aug month typo; render midnight
  as 12 AM and noon as 12 PM
- audio-alsa: bare DIE() now returns -1 instead of falling through
- remove duplicate webcam.h include from blocks.def.h

tests:
- stress: send unused rt signal as child noise instead of SIGUSR1,
  which the binary does not handle; it raced sigterm delivery and
  failed intermittently (now stable over repeated runs)

build:
- Makefile: auto-generate header deps via -MMD -MP; temp.o/procfs.o/
  shell.o previously had no header deps and edits relinked stale objects
- Makefile: regenerate bin scripts when config.h changes
- Makefile: clean leftover test binaries; complete .PHONY

docs:
- AGENTS.md: performance changes must ship with benchmark evidence

+++ blocks/disk.c
@@ -104,7 +104,7 @@ b_write_disk_usage_percent(char *dst, unsigned int dst_size, const char *mountpo
 	if (unlikely(usage == (unsigned int)-1))
 		DIE(return NULL);
 	char *p = dst;
-	p = u_ulltoa_p((unsigned int)usage, p);
+	p = u_ulltoa_p(usage, p);
 	return p;
 	(void)dst_size;
 	(void)interval;
